requirements.txt
# Avalanche framework for continual learning
avalanche-lib==0.5.0

# Deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0

# Vision models
timm>=0.9.0  # For various backbone models

# Experiment tracking and visualization
tensorboard>=2.11.0
wandb>=0.15.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Configuration and utilities
hydra-core>=1.3.0
omegaconf>=2.3.0
pandas>=1.5.0
numpy>=1.23.0
tqdm>=4.65.0

# Evaluation metrics
scikit-learn>=1.2.0